Rename servce -> server
